+++ src/bin/vip-dev-env-create.js
@@ -143,7 +143,7 @@ cmd.argv( process.argv, async ( arg, opt ) => {
 	try {
 		await createEnvironment( instanceData );
 
-		await printEnvironmentInfo( lando, slug, { extended: false } );
+		await printEnvironmentInfo( lando, slug, { extended: false, suppressWarnings: true } );
 
 		const message = '\n' + chalk.green( '✓' ) + ` environment created.\n\nTo start it please run:\n\n${ startCommand }\n`;
 		console.log( message );