+++ src/lib/dev-environment/dev-environment-cli.js
@@ -20,6 +20,7 @@ import dns from 'dns';
  */
 import { ProgressTracker } from '../../lib/cli/progress';
 import { trackEvent } from '../tracker';
+
 import {
 	DEV_ENVIRONMENT_FULL_COMMAND,
 	DEV_ENVIRONMENT_DEFAULTS,
@@ -300,7 +301,7 @@ export async function promptForArguments( preselectedOptions: InstanceOptions, d
 		multisite: preselectedOptions.multisite !== undefined ? preselectedOptions.multisite : await promptForBoolean( multisiteText, !! multisiteDefault ),
 		elasticsearch: false,
 		php: preselectedOptions.php ? resolvePhpVersion( preselectedOptions.php ) : await promptForPhpVersion( resolvePhpVersion( defaultOptions.php || DEV_ENVIRONMENT_DEFAULTS.phpVersion ) ),
-		mariadb: preselectedOptions.mariadb || defaultOptions.mariadb || DEV_ENVIRONMENT_DEFAULTS.mariadbVersion,
+		mariadb: preselectedOptions.mariadb || defaultOptions.mariadb,
 		mediaRedirectDomain: preselectedOptions.mediaRedirectDomain || '',
 		wordpress: {
 			mode: 'image',
@@ -377,7 +378,7 @@ async function processComponent( component: string, preselectedValue: string, de
 	if ( preselectedValue ) {
 		result = processComponentOptionInput( preselectedValue, allowLocal );
 
-		if ( suppressPrompts === false ) {
+		if ( allowLocal && suppressPrompts === false ) {
 			console.log( `${ chalk.green( '✓' ) } Path to your local ${ componentDisplayNames[ component ] }: ${ preselectedValue }` );
 		}
 	} else {
@@ -500,8 +501,7 @@ function resolvePhpVersion( version: string ): string {
 	const versions = Object.keys( DEV_ENVIRONMENT_PHP_VERSIONS );
 	const images = ( ( Object.values( DEV_ENVIRONMENT_PHP_VERSIONS ): any[] ): string[] );
 
-	// eslint-disable-next-line eqeqeq -- use loose comparison because commander resolves '8.0' to '8'
-	const index = versions.findIndex( value => value == version );
+	const index = versions.findIndex( value => value === version );
 	if ( index === -1 ) {
 		const image = images.find( value => value === version );
 		return image ?? images[ 0 ];
@@ -618,18 +618,27 @@ export function processBooleanOption( value: string ): boolean {
 	return ! ( FALSE_OPTIONS.includes( value.toLowerCase?.() ) );
 }
 
+export function processVersionOption( value: string ): string {
+	if ( ! isNaN( value ) && value % 1 === 0 ) {
+		// If it's an Integer passed in, let's ensure that it has a decimal in it to match the version tags e.g. 6 => 6.0
+		return parseFloat( value ).toFixed( 1 );
+	}
+
+	return value;
+}
+
 export function addDevEnvConfigurationOptions( command: Command ): any {
+	// We leave the third parameter to undefined on some because the defaults are handled in preProcessInstanceData()
 	return command
-		.option( 'wordpress', 'Use a specific WordPress version' )
+		.option( 'wordpress', 'Use a specific WordPress version', undefined, processVersionOption )
 		.option( [ 'u', 'mu-plugins' ], 'Use a specific mu-plugins changeset or local directory' )
 		.option( 'app-code', 'Use the application code from a local directory or use "demo" for VIP skeleton code' )
 		.option( 'phpmyadmin', 'Enable PHPMyAdmin component. By default it is disabled', undefined, processBooleanOption )
 		.option( 'xdebug', 'Enable XDebug. By default it is disabled', undefined, processBooleanOption )
 		.option( 'xdebug_config', 'Extra configuration to pass to xdebug via XDEBUG_CONFIG environment variable' )
 		.option( 'elasticsearch', 'Enable Elasticsearch (needed by Enterprise Search)', undefined, processBooleanOption )
-		.option( 'mariadb', 'Explicitly choose MariaDB version to use' )
 		.option( [ 'r', 'media-redirect-domain' ], 'Domain to redirect for missing media files. This can be used to still have images without the need to import them locally.' )
-		.option( 'php', 'Explicitly choose PHP version to use' )
+		.option( 'php', 'Explicitly choose PHP version to use', undefined, processVersionOption )
 		.option( [ 'A', 'mailhog' ], 'Enable MailHog. By default it is disabled', undefined, processBooleanOption );
 }
 
@@ -640,22 +649,22 @@ export async function getTagChoices(): Promise<{ name: string, message: string, 
 	let versions = await getVersionList();
 	if ( versions.length < 1 ) {
 		versions = [ {
+			ref: '6.1.1',
+			tag: '6.1',
+			cacheable: true,
+			locked: true,
+			prerelease: false,
+		},
+		{
+			ref: '6.0.3',
+			tag: '6.0',
+			cacheable: true,
+			locked: true,
+			prerelease: false,
+		},
+		{
 			ref: '5.9.5',
 			tag: '5.9',
-			cacheable: true,
-			locked: true,
-			prerelease: false,
-		},
-		{
-			ref: '5.8.6',
-			tag: '5.8',
-			cacheable: true,
-			locked: true,
-			prerelease: false,
-		},
-		{
-			ref: '5.7.8',
-			tag: '5.7',
 			cacheable: true,
 			locked: true,
 			prerelease: false,