+++ src/lib/utils.js
@@ -1,4 +1,23 @@
 /**
+ * @flow
+ * @format
+ */
+
+/**
+ * External dependencies
+ */
+import fs from 'fs';
+import path from 'path';
+import os from 'os';
+import debugLib from 'debug';
+
+/**
+ * Internal dependencies
+ */
+
+const debug = debugLib( '@automattic/vip:lib:utils' );
+
+/**
  * Polls a function until its return value satisfies a condition
  *
  * @param {Function} fn       A function to poll
@@ -19,4 +38,23 @@ export async function pollUntil( fn, interval, isDone ) {
 		// eslint-disable-next-line no-await-in-loop
 		await new Promise( res => setTimeout( res, interval ) );
 	}
+}
+
+/**
+ * Create a temporary directory in the system's temp directory
+ *
+ * @param {string} prefix Prefix for the directory name
+ * @return {string}				Path to the temporary directory
+ * @throws {Error} 				If the directory cannot be created
+ */
+export function makeTempDir( prefix = 'vip-cli' ) {
+	const tempDir = fs.mkdtempSync( path.join( os.tmpdir(), `${ prefix }-` ) );
+	debug( `Created a directory to hold temporary files: ${ tempDir }` );
+
+	process.on( 'exit', () => {
+		fs.rmSync( tempDir, { recursive: true, force: true } );
+		debug( `Removed temporary directory: ${ tempDir }` );
+	} );
+
+	return tempDir;
 }